💾📂 Database Interaction Refactor
