+++ file-server/internal/helpers/token_helpers.go
@@ -1,0 +1,22 @@
+package helpers
+
+import (
+	"crypto/rand"
+	"crypto/sha256"
+	"encoding/hex"
+
+)
+
+func HashPassword(password, salt string) string {
+	hash := sha256.Sum256([]byte(password + salt))
+	return hex.EncodeToString(hash[:])
+}
+
+func GenerateRandomSalt() (string, error) {
+	salt := make([]byte, 16)
+	_, err := rand.Read(salt)
+	if err != nil {
+		return "", err
+	}
+	return hex.EncodeToString(salt), nil
+}
